requirements_agent.txt
fastapi==0.115.0
uvicorn==0.30.6
google-genai==1.3.0
pandas==2.2.3
python-dotenv==1.0.1
pydantic==2.10.1
sqlalchemy==2.0.36
jinja2==3.1.4
python-multipart==0.0.18
openpyxl==3.1.5